+++ crate/cli/src/actions/abe/decrypt.rs
@@ -1,6 +1,6 @@
 use std::{fs::File, io::prelude::*, path::PathBuf};
 
-use clap::StructOpt;
+use clap::Parser;
 use cosmian_kmip::kmip::kmip_operations::DecryptedData;
 use cosmian_kms_client::KmsRestClient;
 use cosmian_kms_utils::crypto::generic::kmip_requests::build_decryption_request;
@@ -8,22 +8,22 @@ use eyre::Context;
 
 /// Decrypts a file identified by its name and
 /// given a user decryption key stored in the KMS.
-#[derive(StructOpt, Debug)]
+#[derive(Parser, Debug)]
 pub struct DecryptAction {
     /// The file to decrypt
-    #[structopt(required = true, name = "FILE", parse(from_os_str))]
+    #[clap(required = true, name = "FILE")]
     input_file: PathBuf,
 
     /// The encrypted output file path
-    #[structopt(required = false, parse(from_os_str), long, short = 'o')]
+    #[clap(required = false, long, short = 'o')]
     output_file: PathBuf,
 
     /// The optional resource_uid. It's an extra encryption parameter to increase the security level
-    #[structopt(required = false, long, short, default_value = "")]
+    #[clap(required = false, long, short, default_value = "")]
     resource_uid: String,
 
     /// The user decryption key unique identifier stored in the KMS
-    #[structopt(required = true, long, short = 'u')]
+    #[clap(required = true, long, short = 'u')]
     user_key_id: String,
 }