+++ crate/server/src/database/pgsql.rs
@@ -653,7 +653,7 @@ impl Database for Pgsql {
 //
 #[cfg(test)]
 mod tests {
-    use cosmian_crypto_core::CsRng;
+    use cosmian_crypto_core::{reexport::rand_core::SeedableRng, CsRng};
     use cosmian_kmip::kmip::{
         kmip_objects::ObjectType,
         kmip_types::{
@@ -662,7 +662,6 @@ mod tests {
         },
     };
     use cosmian_kms_utils::{crypto::aes::create_symmetric_key, types::ObjectOperationTypes};
-    use rand_core::SeedableRng;
     use serial_test::serial;
     use uuid::Uuid;