+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+---
+
+## [4.3.3] - 2023-03-02
+
+### Features
+
+- Use CoverCrypt v11
+
+### Refactor
+
+- Removed `mysql` crate used for EdgelessDB (compatible with client SSL connection)
+- Use workspace dependencies to ease maintenance
+
+---
+
 ## [4.3.2] - 2023-02-17
 
 ### Documentation